Add ``estimate_qpu_access_time`` method: implemented most review comments on mock restructure
